requirements.txt
tensorflow
torch
torchvision
timm
librosa
kagglehub
scikit-learn
matplotlib
pandas
numpy
opencv-python
pillow
soundfile
